refactor: extract super ops security sessions
